test_from_stori_card/requirements.txt
asgiref==3.7.2
astroid==3.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
platformdirs==4.2.0
pluggy==1.4.0
psycopg2==2.9.9
pylint==3.1.0
pytest==8.0.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
